Factor out fuzzy date formatting
